pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TextArena"
version = "0.7.3"
description = "A Collection of Competitive Text-Based Games for Language Model Evaluation and Reinforcement Learning"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="Leon Guertler", email="Guertlerlo@cfar.a-star.edu.sg"},]
dependencies = [
    "openai",
    "rich",
    "nltk",
    "chess",
    "python-dotenv",
    "requests",
    "websockets",
    "google-genai>=1.34.0",
    "pyyaml>=6.0.2",
]

[project.urls]
"Homepage" = "https://www.textarena.ai/"

[project.optional-dependencies]
all = ["sympy", "latex2sympy", "openai", "google-generativeai", "transformers", "cerebras-cloud-sdk", "boto3", "anthropic"]


